feat: play segment on click quranword
